+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/inlang.config.js
