+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/+layout.server.ts
@@ -1,5 +1,7 @@
 import type { LayoutServerLoad } from "./$types.js"
 
+export const prerender = true
+
 export const load = (async ({ locals, url }) => {
 	console.info("+layout.server.ts", locals.i18n.i("welcome"))
