+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/+page.server.ts
