+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/src/routes/inlang/[language]/+server.ts
@@ -1,6 +1,6 @@
 import { json } from "@sveltejs/kit"
 import type { RequestHandler } from "./$types.js"
-import { getResource } from "../../../inlang.server.js"
+import { getResource } from "../../../inlang.server"
 
 // eslint-disable-next-line unicorn/no-null
 export const GET = (({ params: { language } }) =>
