+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/static/favicon.ico
