+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/svelte.config.js
@@ -1,4 +1,5 @@
-import adapter from "@sveltejs/adapter-auto"
+// import adapter from "@sveltejs/adapter-static"
+import adapter from "@sveltejs/adapter-node"
 import { vitePreprocess } from "@sveltejs/kit/vite"
 
 /** @type {import('@sveltejs/kit').Config} */
@@ -12,6 +13,9 @@ const config = {
 		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter(),
+		prerender: {
+			entries: ["*"],
+		},
 	},
 }
