+++ source-code/sdk-js/src/adapter-sveltekit/examples/lang-in-slug/tsconfig.json
